# docs/api-documentation-openapi.json
{
    "openapi": "3.0.1",
...
    "info": {
        "title": "mimoto",
        "description": "mimoto",
        "license": {
            "name": "Mosip",
            "url": "https://docs.mosip.io/platform/license"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://api.collab.mosip.net/v1/mimoto",
            "description": "For Swagger"
        }
    ],
    "paths": {
...
        "/allProperties": {
            "get": {
                "operationId": "getAllProperties",
                "responses": {},
                "description": "get all configurable properties for inji.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "id": {
                                        "type": "string"
                                    },
                                    "version": {
                                        "type": "string"
                                    },
                                    "str": {
                                        "type": "null"
                                    },
                                    "responsetime": {
                                        "type": "string"
                                    },
                                    "metadata": {
                                        "type": "null"
                                    },
                                    "response": {
                                        "type": "object",
                                        "properties": {
                                            "internal.auth.types.allowed": {
                                                "type": "string"
                                            },
                                            "auth.types.allowed": {
                                                "type": "string"
                                            },
                                            "ekyc.auth.types.allowed": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "errors": {
                                        "type": "array",
                                        "items": {
                                            "type": "object",
                                            "properties": {}
                                        }
                                    }
                                },
                                "x-examples": {
                                    "Example 1": {
                                        "id": "mosip.inji.properties",
                                        "version": "v1",
                                        "str": null,
                                        "responsetime": "2022-10-31T05:22:21.009Z",
                                        "metadata": null,
                                        "response": {
                                            "internal.auth.types.allowed": "otp,bio-Finger,bio-Iris,bio-Face",
                                            "auth.types.allowed": "demo,otp,bio-Finger,bio-Iris,bio-Face",
                                            "ekyc.auth.types.allowed": "demo,otp,bio-Finger,bio-Iris,bio-Face"
                                        },
                                        "errors": []
                                    }
                                }
                            },
                            "examples": {
                                "example-1": {
                                    "value": {
                                        "id": "mosip.inji.properties",
                                        "version": "v1",
                                        "str": null,
                                        "responsetime": "2022-10-31T05:22:21.009Z",
                                        "metadata": null,
                                        "response": {
                                            "internal.auth.types.allowed": "otp,bio-Finger,bio-Iris,bio-Face",
                                            "auth.types.allowed": "demo,otp,bio-Finger,bio-Iris,bio-Face",
                                            "ekyc.auth.types.allowed": "demo,otp,bio-Finger,bio-Iris,bio-Face"
                                        },
                                        "errors": []
                                    }
                                }
                            }
                        }
                    }
                },
                "summary": "Get all Inji app properties"
            }
        },
...
    },
...
}